feat: refactor event handling for snippets panel, update component imports, and remove unused files
