+++ src/app/dashboard/page.tsx
@@ -1,6 +1,6 @@
-import {AppSidebar} from "@/components/app-sidebar";
-import EditorColumn from "@/components/editor-column";
-import SnippetsPanel from "@/components/snippets-panel";
+import {AppSidebar} from "@/components/sidebar/app-sidebar";
+import EditorColumn from "@/components/editor-panel/editor-column";
+import SnippetsPanel from "@/components/snippets-panel/snippets-panel";
 import {ResizableHandle, ResizablePanelGroup} from "@/components/ui/resizable";
 import {SidebarProvider} from "@/components/ui/sidebar";
 import {getCollections} from "@/lib/db/data/get_collections";
