+++ src/components/editor-panel/editor-footer.tsx
