+++ src/components/snippets-panel/snippet-card.tsx
