+++ src/components/snippets-panel/snippets-list.tsx
