+++ src/components/snippets-panel/snippets-panel.tsx
@@ -4,16 +4,17 @@ import React from "react";
 import {Search} from "lucide-react";
 import {useQuery} from "@tanstack/react-query";
 
-import SnippetsPanelContainer from "./snippets-panel-container";
-import {SidebarTrigger} from "./ui/sidebar";
-import Settings from "./settings";
-import {ModeToggle} from "./theme-toggle";
+import {SidebarTrigger} from "../ui/sidebar";
+import Settings from "../editor-panel/settings";
+import {ModeToggle} from "../layout/theme-toggle";
+import {Separator} from "../ui/separator";
+import {ScrollArea} from "../ui/scroll-area";
+import {Input} from "../ui/input";
+import {CreateSnippetForm} from "../forms/create-snippet-form";
+import {Skeleton} from "../ui/skeleton";
+
 import SnippetsLits from "./snippets-list";
-import {Separator} from "./ui/separator";
-import {ScrollArea} from "./ui/scroll-area";
-import {Input} from "./ui/input";
-import {CreateSnippetForm} from "./forms/create-snippet-form";
-import {Skeleton} from "./ui/skeleton";
+import SnippetsPanelContainer from "./snippets-panel-container";
 
 import {getFolderAndSnippetsById} from "@/lib/db/data/get_folder_and_snippets";
 import {cn} from "@/lib/utils";
