--- src/lib/colletions-mock-data/mock-data-users.ts
@@ -1,16 +1,0 @@
-export const users = [
-  {
-    id: "u123a456b-789c-def0-123a-bc45def67890",
-    username: "devuser",
-    email: "dev@example.com",
-    createdAt: new Date("2025-01-10T08:00:00Z"),
-    updatedAt: new Date("2025-01-10T08:00:00Z"),
-  },
-  {
-    id: "u987z654y-321x-wvu0-987z-yx65wvu43210",
-    username: "anotheruser",
-    email: "another@example.com",
-    createdAt: new Date("2025-01-12T10:15:00Z"),
-    updatedAt: new Date("2025-01-12T10:15:00Z"),
-  },
-];
