--- src/lib/constants/snippet-options.ts
@@ -1,67 +1,0 @@
-export const snippetsOptions = {
-  "programming-language": [
-    {value: "javascript", label: "JavaScript"},
-    {value: "typescript", label: "TypeScript"},
-    {value: "python", label: "Python"},
-    {value: "rust", label: "Rust"}, //
-    {value: "go", label: "Go"},
-    {value: "kotlin", label: "Kotlin"},
-    {value: "swift", label: "Swift"},
-    {value: "java", label: "Java"},
-    {value: "csharp", label: "C#"},
-    {value: "php", label: "PHP"},
-    {value: "ruby", label: "Ruby"},
-    {value: "sql", label: "SQL"},
-  ],
-  framework: [
-    {value: "react", label: "React"},
-    {value: "nextjs", label: "Next.js"},
-    {value: "vue", label: "Vue"},
-    {value: "nuxtjs", label: "Nuxt.js"},
-    {value: "sveltekit", label: "SvelteKit"},
-    {value: "angular", label: "Angular"},
-
-    {value: "nestjs", label: "NestJS"},
-    {value: "express", label: "Express"},
-    {value: "laravel", label: "Laravel"},
-    {value: "springboot", label: "Spring Boot"},
-    {value: "django", label: "Django"},
-    {value: "flask", label: "Flask"},
-    {value: "remix", label: "Remix"},
-    {value: "dotnet", label: ".NET"},
-  ],
-  tool: [
-    {value: "docker", label: "Docker"},
-    {value: "kubernetes", label: "Kubernetes"},
-    {value: "github-actions", label: "GitHub Actions"},
-    {value: "terraform", label: "Terraform"},
-
-    {value: "vite", label: "Vite"},
-    {value: "turbo", label: "TurboRepo"},
-    {value: "eslint", label: "ESLint"},
-    {value: "prettier", label: "Prettier"},
-
-    {value: "jest", label: "Jest"},
-    {value: "cypress", label: "Cypress"},
-    {value: "playwright", label: "Playwright"},
-
-    {value: "pnpm", label: "pnpm"},
-    {value: "bun", label: "Bun"},
-
-    {value: "aws", label: "AWS"},
-    {value: "azure", label: "Azure"},
-    {value: "gcp", label: "Google Cloud"},
-
-    {value: "git", label: "Git"},
-    {value: "postman", label: "Postman"},
-    {value: "tailwindcss", label: "Tailwind CSS"},
-  ],
-  "markup-styling": [
-    {value: "html", label: "HTML"},
-    {value: "css", label: "CSS"},
-    {value: "sass", label: "Sass"},
-    {value: "markdown", label: "Markdown"},
-    {value: "jsx", label: "JSX"},
-    {value: "tsx", label: "TSX"},
-  ],
-};
