+++ src/lib/db/actions/snippets/update-snippet-language.tsx
@@ -1,17 +1,15 @@
 "use server";
 
-import {revalidatePath} from "next/cache";
-
 import {auth} from "@/lib/auth";
 import db from "@/lib/db";
 import {Language} from "@/types";
 
 export async function updateSnippetLanguage(snippetId: string, language: Language) {
   try {
-    const user = await auth();
+    const session = await auth();
 
-    if (!user) {
-      throw new Error("User is not authenticated");
+    if (!session?.user?.id) {
+      throw new Error("Unauthorized");
     }
 
     const snippet = await db.snippet.update({
@@ -25,8 +23,6 @@ export async function updateSnippetLanguage(snippetId: string, language: Languag
 
     return snippet;
   } catch (error) {
-    throw new Error(`Failed to update snippet language: ${error as string}`);
-  } finally {
-    revalidatePath("/dashboard");
+    throw new Error(error instanceof Error ? error.message : "Error updating snippet language");
   }
 }
