+++ src/lib/utils.ts
@@ -7,8 +7,10 @@ const cn = (...inputs: ClassValue[]) => {
 
 export {cn};
 
-export function capitalize(str: string): string {
+function capitalize(str: string): string {
   if (!str) return "";
 
   return str.charAt(0).toUpperCase() + str.slice(1);
 }
+
+export {capitalize};
